+++ src/swarm_rescue/maps/map_complete_02.py
@@ -1,6 +1,8 @@
 import math
 import random
+from typing import List
 
+from spg.playground import Playground
 from spg.utils.definitions import CollisionTypes
 
 from spg_overlay.gui_map.closed_playground import ClosedPlayground
@@ -23,13 +25,13 @@ class MyMapComplete02(MapAbstract):
         self._number_drones = 10
         self._time_step_limit = 1200
         self._real_time_limit = 240  # In seconds
-        self._number_wounded_persons = 0  # it will be filled in the function build_map()
+        self._number_wounded_persons = 0  # it will be filled in the function construct_playground()
 
         # BUILD MAP
         self._size_area = (1113, 750)
-        self._wounded_persons = list()
+        self._wounded_persons: List[WoundedPerson] = []
 
-    def construct_playground(self):
+    def construct_playground(self) -> Playground:
         playground = ClosedPlayground(size=self._size_area)
 
         # RESCUE CENTER