+++ src/swarm_rescue/maps/map_random.py
@@ -1,6 +1,10 @@
 import math
 import random
+from typing import List
 
+from spg.playground import Playground
+
+from spg_overlay.entities.wounded_person import WoundedPerson
 from spg_overlay.gui_map.closed_playground import ClosedPlayground
 from spg_overlay.entities.sensor_disablers import EnvironmentType
 from spg_overlay.gui_map.map_abstract import MapAbstract
@@ -15,8 +19,9 @@ class MyMapRandom(MapAbstract):
         self._time_step_limit = 480
         self._real_time_limit = 22  # In seconds
         self._size_area = (1500, 700)
+        self._wounded_persons: List[WoundedPerson] = []
 
-    def construct_playground(self):
+    def construct_playground(self) -> Playground:
         playground = ClosedPlayground(size=self._size_area)
 
         self._explored_map.initialize_walls(playground)