+++ src/swarm_rescue/solutions/my_drone_random.py
@@ -69,8 +69,8 @@ class MyDroneRandom(DroneAbstract):
             self.angleStopTurning = random.uniform(-math.pi, math.pi)
 
         measured_angle = 0
-        if self.measured_angle() is not None:
-            measured_angle = self.measured_angle()
+        if self.measured_compass_angle() is not None:
+            measured_angle = self.measured_compass_angle()
 
         diff_angle = normalize_angle(self.angleStopTurning - measured_angle)
         if self.isTurning and abs(diff_angle) < 0.2: